refactor: change QucikMenu button directory location
